// src/types/bindings.ts
export type SeasonalAnime = {
	monday?: string[],
	tuesday?: string[],
	wednesday?: string[],
	thursday?: string[],
	friday?: string[],
	saturday?: string[],
	sunday?: string[],
};

export type SortingMethod = "release_date_desc" | "release_date_asc" | "title_asc" | "title_desc" | "recently_added" | "oldest_added" | "recently_modified" | "oldest_modified";

export type UpdateSettingsPayload = {
	otakudesu_url: string | null,
	anime_directory: string | null,
	youtube_download_directory: string | null,
	seasonal_anime: SeasonalAnime | null,
};